+++ src/CHANGES.md
@@ -1,8 +1,8 @@
 **Added:**
-* display properties editor
+* export header file types option in export parameters 
 
 **Changed:**
-* how a set macro is displayed in ControlPropertiesEditorPane (centered the label and made the macro key a hyperlink so that when it's clicked, opens macro editor popup)
+* none
 
 **Fixed:**
 * none